The experiement key is now stores as a long instead of an int. It's given back as a Long from the Service Layer.

+++ org.sopeco.webui.shared/src/org/sopeco/webui/shared/entities/ScenarioDetails.java
@@ -64,7 +64,7 @@ public class ScenarioDetails implements Serializable {
 
 	private String selectedExperiment;
 	
-	private int experimentKey;
+	private long experimentKey;
 
 	/**
 	 * Empty constructor for JSON converting.
@@ -181,7 +181,7 @@ public class ScenarioDetails implements Serializable {
 	/**
 	 * @return the experimentKey
 	 */
-	public int getExperimentKey() {
+	public long getExperimentKey() {
 		return experimentKey;
 	}
 
@@ -189,7 +189,7 @@ public class ScenarioDetails implements Serializable {
 	 * @param experimentKey
 	 *            the epxeriment key to set
 	 */
-	public void setExperimentKey(int experimentKey) {
+	public void setExperimentKey(long experimentKey) {
 		this.experimentKey = experimentKey;
 	}
 }
